grpc-ctf/requirements.txt
flask==3.1.2
grpcio-tools==1.75.1
protobuf==6.33.0
grpcio==1.75.1
